6/23 -6

+++ sketch_main.js
@@ -250,7 +250,7 @@ if (windowWidth <= windowSm) {
 
 
         //if (p.frameCount < 15) {
-          if (int(p.frameCount / 5) == 0) {
+          if ((p.frameCount / 5) == 0) {
 
 
               p.noFill();
